feat: profile page code refactoring
