+++ src/pages/profile-page/index.module.css
@@ -5,10 +5,6 @@
     padding-top: 6em;
 }
 
-.main {
-    border-top: 1px solid lightgray;
-}
-
 .header {
     display: flex;
     margin-top: 30px;
@@ -24,77 +20,7 @@
     padding-bottom: 1em;
 }
 
-.section div h2 {
-    display: block;
-    color: rgba(var(--i1d, 38, 38, 38), 1);
-    font-weight: 300;
-    font-size: 28px;
-    line-height: 32px;
-    margin: -5px 0 -6px;
-    margin-right: 1em;
-}
-
-.button {
-    margin-right: 1em;
-    background-color: transparent;
-    border: 1px solid #dbdbdb;
-    border: 1px solid rgba(var(--ca6, 219, 219, 219), 1);
-    color: rgba(var(--f75, 38, 38, 38), 1);
-    border-radius: 4px;
-    position: relative;
-    text-decoration: none;
-    cursor: pointer;
-    display: block;
-    font-weight: 600;
-    padding: 5px 9px;
-}
-
-.section div h1 {
-    display: inline;
-    font-weight: 600;
-    font-size: 16px;
-    margin: 0;
-}
-
-.section ul {
-    padding: 0;
-}
-
-.section ul li {
-    display: inline-block;
-    margin-right: 3em;
-}
-
-.profilePicDiv {
-    width: 150px;
-    height: 150px;
-    margin: 0 5em;
-}
-
-.profilePic {
-    width: 100%;
-    height: 100%;
-    cursor: pointer;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.empty {
-    text-align: center;
-    padding-top: 2em;
-}
-
-.followBtn {
-    margin-right: 1em;
-    background-color: #0095f6;
-    border: 1px solid #dbdbdb;
-    border: 1px solid rgba(var(--ca6, 219, 219, 219), 1);
-    color: #fff;
-    border-radius: 4px;
-    position: relative;
-    text-decoration: none;
-    cursor: pointer;
-    display: block;
-    font-weight: 600;
-    padding: 8px 24px;
+.description {
+    word-break: break-word;
+    margin-top: 1em;
 }
