+++ src/pages/profile-page/posts/index.module.css
@@ -1,0 +1,9 @@
+.main {
+    border-top: 1px solid lightgray;
+    min-height: 43vh;
+}
+
+.empty {
+    text-align: center;
+    padding-top: 2em;
+}
