+++ src/pages/profile-page/profile-followers/index.module.css
@@ -1,0 +1,4 @@
+.container li {
+    display: inline-block;
+    margin-right: 3em;
+}
