+++ src/pages/profile-page/profile-picture/index.module.css
@@ -1,0 +1,13 @@
+.container {
+    width: 150px;
+    height: 150px;
+    margin: 0 5em;
+}
+
+.container img {
+    width: 100%;
+    height: 100%;
+    cursor: pointer;
+    border-radius: 50%;
+    object-fit: cover;
+}
